% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCosts-package.R
\docType{package}
\name{CohortCosts-package}
\alias{CohortCosts}
\alias{CohortCosts-package}
\title{CohortCosts: Direct Medical Cost Extraction and Health Economics Costing on OMOP CDM}
\description{
Links polymorphic OMOP COST records to clinical events and tracks medical expenditures, unit cost tariffs, and healthcare price indices for OMOP CDM cohorts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.iomed.health/}
  \item \url{https://github.com/iomedhealth/hermes}
}

}
\author{
\strong{Maintainer}: Gabriel Maeztu \email{gabriel@iomed.health}

Authors:
\itemize{
  \item Gabriel Maeztu \email{gabriel@iomed.health}
}

}
\keyword{internal}
